refactor: unify performance pages
