--- src/public/pages/authorized-user/subject-performance/subject-performance.css
@@ -1,52 +1,0 @@
-#chartContainer {
-    background: white;
-    border-radius: 8px;
-    box-shadow: 0 2px 10px rgba(0,0,0,0.1);
-    padding: 20px;
-    margin-bottom: 20px;
-}
-
-.filters {
-    display: flex;
-    gap: 20px;
-    margin-bottom: 20px;
-    flex-wrap: wrap;
-    justify-content: center;
-}
-
-.filters label {
-    display: flex;
-    flex-direction: column;
-    font-weight: bold;
-    gap: 5px;
-}
-
-.filters select {
-    padding: 8px 12px;
-    border-radius: 4px;
-    border: 1px solid #ccc;
-    font-size: 16px;
-    min-width: 150px;
-}
-
-h1 {
-    color: #0B2072;
-    margin-bottom: 20px;
-}
-
-#suggestionBox h3, #medalsBox h3 {
-    color: #3D52A0;
-    border-bottom: 2px solid #3D52A0;
-    padding-bottom: 5px;
-}
-
-@media (max-width: 768px) {
-    .bottom-container {
-        flex-direction: column;
-    }
-    
-    #chartContainer {
-        width: 95%;
-        padding: 10px;
-    }
-}
