+++ src/public/pages/not-authorized-user/script.js
@@ -9,7 +9,6 @@ import LeaderboardPage from "../authorized-user/leaderboard-page/leaderboard-pag
 import ExamPage from "../authorized-user/exam/exam.js";
 import SearchPage from "../authorized-user/search/search.js";
 import HistoryPage from "../authorized-user/exam-history/exam-history.js";
-import SubjectPerformancePage from "../authorized-user/subject-performance/subject-performance.js";
 import PerformancePage from "../authorized-user/performance/performance.js";
 
 const app = document.getElementById("app");
@@ -23,8 +22,6 @@ const stylePaths = {
   search: "/pages/authorized-user/search/search.css",
   profile: "/pages/authorized-user/user-profile/user.css",
   exam: "/pages/authorized-user/exam/exam.css",
-  subjectPerformance:
-    "/pages/authorized-user/subject-performance/subject-performance.css",
   performance: "/pages/authorized-user/performance/performance.css",
 };
 
@@ -77,11 +74,6 @@ const routes = {
   "/exam": {
     component: () => ExamPage(),
     style: stylePaths.exam,
-    requiresAuth: true,
-  },
-  "/subject-performance": {
-    component: () => SubjectPerformancePage(),
-    style: stylePaths.subjectPerformance,
     requiresAuth: true,
   },
   "/performance": {
